Rename these files since "bgg" has moved into the directory structure.
