+++ lib/bgg.rb
@@ -41,13 +41,13 @@ class BggApi
 end
 
 
-require 'bgg/bgg_collection'
-require 'bgg/bgg_collection_item'
-require 'bgg/bgg_game'
-require 'bgg/bgg_play'
-require 'bgg/bgg_plays'
-require 'bgg/bgg_plays_iterator'
-require 'bgg/bgg_search'
-require 'bgg/bgg_search_result'
-require 'bgg/bgg_user'
+require 'bgg/collection'
+require 'bgg/collection_item'
+require 'bgg/game'
+require 'bgg/play'
+require 'bgg/plays'
+require 'bgg/plays_iterator'
+require 'bgg/search'
+require 'bgg/search_result'
+require 'bgg/user'
