+++ lib/bgg/collection.rb
