+++ lib/bgg/collection_item.rb
