+++ lib/bgg/game.rb
