+++ lib/bgg/play.rb
