+++ lib/bgg/plays_iterator.rb
