+++ lib/bgg/search.rb
