+++ lib/bgg/search_result.rb
