+++ lib/bgg/user.rb
